// migrations/snapshots/97d8d2efdb742858bf44d6a35996a98d6902373ede663eee9e2baabef6599cdb/contract.d.ts
export type CodecTypes = MongoCodecTypes;

export type FieldOutputTypes = {
  readonly __unbound__: {
    readonly Chats: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly role: CodecTypes['mongo/string@1']['output'];
      readonly content: CodecTypes['mongo/string@1']['output'];
      readonly chatSessionId: CodecTypes['mongo/objectId@1']['output'];
      readonly createdAt: CodecTypes['mongo/date@1']['output'];
      readonly updatedAt: CodecTypes['mongo/date@1']['output'];
    };
    readonly ChatSession: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly name: CodecTypes['mongo/string@1']['output'];
      readonly userId: CodecTypes['mongo/objectId@1']['output'];
      readonly createdAt: CodecTypes['mongo/date@1']['output'];
      readonly updatedAt: CodecTypes['mongo/date@1']['output'];
    };
    readonly UploadedPdf: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly documentId: CodecTypes['mongo/string@1']['output'];
      readonly key: CodecTypes['mongo/string@1']['output'];
      readonly fileName: CodecTypes['mongo/string@1']['output'];
      readonly userId: CodecTypes['mongo/objectId@1']['output'];
      readonly createdAt: CodecTypes['mongo/date@1']['output'];
      readonly updatedAt: CodecTypes['mongo/date@1']['output'];
    };
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly email: CodecTypes['mongo/string@1']['output'];
      readonly googleId: CodecTypes['mongo/string@1']['output'];
      readonly name: CodecTypes['mongo/string@1']['output'] | null;
      readonly image: CodecTypes['mongo/string@1']['output'] | null;
      readonly createdAt: CodecTypes['mongo/date@1']['output'];
      readonly updatedAt: CodecTypes['mongo/date@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly __unbound__: {
    readonly Chats: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly role: CodecTypes['mongo/string@1']['input'];
      readonly content: CodecTypes['mongo/string@1']['input'];
      readonly chatSessionId: CodecTypes['mongo/objectId@1']['input'];
      readonly createdAt: CodecTypes['mongo/date@1']['input'];
      readonly updatedAt: CodecTypes['mongo/date@1']['input'];
    };
    readonly ChatSession: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly name: CodecTypes['mongo/string@1']['input'];
      readonly userId: CodecTypes['mongo/objectId@1']['input'];
      readonly createdAt: CodecTypes['mongo/date@1']['input'];
      readonly updatedAt: CodecTypes['mongo/date@1']['input'];
    };
    readonly UploadedPdf: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly documentId: CodecTypes['mongo/string@1']['input'];
      readonly key: CodecTypes['mongo/string@1']['input'];
      readonly fileName: CodecTypes['mongo/string@1']['input'];
      readonly userId: CodecTypes['mongo/objectId@1']['input'];
      readonly createdAt: CodecTypes['mongo/date@1']['input'];
      readonly updatedAt: CodecTypes['mongo/date@1']['input'];
    };
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly email: CodecTypes['mongo/string@1']['input'];
      readonly googleId: CodecTypes['mongo/string@1']['input'];
      readonly name: CodecTypes['mongo/string@1']['input'] | null;
      readonly image: CodecTypes['mongo/string@1']['input'] | null;
      readonly createdAt: CodecTypes['mongo/date@1']['input'];
      readonly updatedAt: CodecTypes['mongo/date@1']['input'];
    };
  };
};
// ...
export type TypeMaps = MongoTypeMaps<CodecTypes, FieldOutputTypes, FieldInputTypes>;
// ...
export type Contract = MongoContractWithTypeMaps<ContractBase, TypeMaps>;